Add a repository for saving and loading events
I was really hoping this change would force me to create aggregates with
an ID but instead I realised that I had no way of figuring out an
aggregate's pid once restarted by the supervisor. I could have used this
as an excuse to pass an ID and register the processes with `{:global,
id}` or `{:via, Custom.Registry, id}` BUT that introduces global state
which is probably fine but not what I wanted to do.

Instead I used the technique from the mix-otp getting started guide
(http://elixir-lang.org/getting-started/mix-otp/supervisor-and-application.html)
and wrote my own supervised registry.