Change supervisor strategy to `:rest_for_one`
At some stage I need to re-evaluate if this supervision tree is correct.
There is a supervisor per aggregate type that manages individual
aggregates BUT there is a shared identity map between all aggregates.
If the map fails all aggregate type supervisors (and therefore
aggregates) need to be restarted. Right now I'm OK with this but maybe
an identity map per type is a better approach.